All code working

+++ PHAS3459/src/module6/PowerLawTheory.java
@@ -9,7 +9,7 @@ public class PowerLawTheory implements Theory {
   private double n; // Power to raise x to
 
   // Constructor sets value of n
-  public PowerLawTheory(int n) {
+  public PowerLawTheory(double n) {
     this.n = n;
   }
